+++ omc3/segment_by_segment/propagables/alpha.py
@@ -7,15 +7,17 @@ This module contains the propagable for the alpha parameter.
 """
 from __future__ import annotations
 
-from functools import cache
 from typing import TYPE_CHECKING
 
 import pandas as pd
+from tfs import TfsDataFrame
 
 from omc3.definitions.optics import OpticsMeasurement
 from omc3.optics_measurements.constants import ALPHA
-from omc3.segment_by_segment.propagables.abstract import Propagable 
-from omc3.segment_by_segment.propagables.utils import PropagableColumns as Columns
+from omc3.segment_by_segment import math as sbs_math
+from omc3.segment_by_segment.propagables.abstract import Propagable
+from omc3.segment_by_segment.propagables.phase import Phase 
+from omc3.segment_by_segment.propagables.utils import PropagableColumns, common_indices
 from omc3.segment_by_segment.segments import SegmentDiffs
 from omc3.utils import logging_tools
 
@@ -29,8 +31,8 @@ LOG = logging_tools.get_logger(__name__)
 
 class AlphaPhase(Propagable):
 
-    _init_pattern = "alf{}_{}"
-    columns: Columns = Columns(ALPHA)
+    _init_pattern = "alf{}_{}"  # format(plane, ini/end)
+    columns: PropagableColumns = PropagableColumns(ALPHA)
 
     @classmethod
     def get_at(cls, names: IndexType, meas: OpticsMeasurement, plane: str) -> ValueErrorType:
@@ -38,6 +40,16 @@ class AlphaPhase(Propagable):
         alpha = meas.beta_phase[plane].loc[names, c.column]
         error = meas.beta_phase[plane].loc[names, c.error_column]
         return alpha, error
+    
+    @classmethod
+    def in_measurement(cls, meas: OpticsMeasurement) -> bool:
+        """ Check if the alpha phase is in the measurement data. """
+        try:
+            meas.beta_phase_x
+            meas.beta_phase_y
+        except FileNotFoundError:
+            return False
+        return True
     
     def init_conditions_dict(self):
         # alpha needs to be inverted for backward propagation, i.e. the end-init
@@ -47,29 +59,74 @@ class AlphaPhase(Propagable):
                 init_cond[key] = -value
         return init_cond
 
-    @cache
-    def measured_forward(self, plane):
-        pass
+    def get_segment_observation_points(self, plane: str):
+        """ Return the measurement points for the given plane, that are in the segment. """
+        return common_indices(
+            self.segment_models.forward.index, 
+            self._meas.beta_phase[plane].index
+        )  
 
-    @cache
-    def correction_forward(self, plane):
-        pass
-
-    @cache
-    def expected_forward(self, plane):
-        pass
-
-    @cache
-    def measured_backward(self, plane):
-        pass
-
-    @cache
-    def correction_backward(self, plane):
-        pass
-
-    @cache
-    def expected_backward(self, plane):
-        pass
-    
     def add_differences(self, segment_diffs: SegmentDiffs):
-        pass
+        """ Calculate the differences between the propagated models and the measured values."""
+        dfs = self.get_difference_dataframes()
+        for plane, df in dfs.items():
+            # save to diffs/write to file (if allow_write is set)
+            segment_diffs.alpha_phase[plane] = df
+
+    def _compute_measured(self, 
+            plane: str, 
+            seg_model: TfsDataFrame, 
+            forward: bool
+        ) -> tuple[pd.Series, pd.Series]:
+        """ Compute the alpha-difference between the given segment model and the measured values."""
+        init_condition = self._init_start(plane) if forward else self._init_end(plane)
+
+        # get the measured values
+        names = self.get_segment_observation_points(plane)
+        alpha, err_alpha = self.get_at(names, self._meas, plane)
+
+        # get the propagated values
+        model_alpha = seg_model.loc[names, f"{ALPHA}{plane}"]
+        if not forward:
+            model_alpha = -model_alpha  # alpha needs to be inverted for backward propagation
+        model_phase = Phase.get_segment_phase(seg_model.loc[names, :], plane, forward) 
+
+        # calculate difference
+        alpha_diff = alpha - model_alpha
+
+        # propagate the error
+        propagated_err = sbs_math.propagate_error_alpha(model_alpha, model_phase, init_condition)
+        total_err = sbs_math.quadratic_add(err_alpha, propagated_err)
+        return alpha_diff, total_err
+    
+    def _compute_correction(
+            self,
+            plane: str,
+            seg_model: pd.DataFrame,
+            seg_model_corr: pd.DataFrame,
+            forward: bool,
+        ) -> tuple[pd.Series, pd.Series]:
+        """Compute the alpha differennce between the nominal and the corrected model."""
+        init_condition = self._init_start(plane) if forward else self._init_end(plane)
+
+        model_alpha = seg_model.loc[:, f"{ALPHA}{plane}"]
+        corrected_alpha = seg_model_corr.loc[:, f"{ALPHA}{plane}"]
+        alpha_diff = corrected_alpha - model_alpha
+        if not forward:
+            alpha_diff = -alpha_diff
+        
+        # propagate the error
+        model_phase = Phase.get_segment_phase(seg_model, plane, forward)
+        propagated_err = sbs_math.propagate_error_alpha(corrected_alpha, model_phase, init_condition)
+        return alpha_diff, propagated_err
+
+    def _compute_elements(self, plane: str, seg_model: pd.DataFrame, forward: bool):
+        """ Compute get the propagated alpha values from the segment model and calculate the propagated error.  """
+        init_condition = self._init_start(plane) if forward else self._init_end(plane)
+
+        model_alpha = seg_model.loc[:, f"{ALPHA}{plane}"]
+
+        # propagate the error
+        model_phase = Phase.get_segment_phase(seg_model, plane, forward)
+        propagated_err = sbs_math.propagate_error_alpha(model_alpha, model_phase, init_condition)
+        return model_alpha, propagated_err