+++ omc3/segment_by_segment/propagables/dispersion.py
@@ -1,9 +1,9 @@
 
 """
-Beta Propagable
----------------
+Dispersion Propagable
+---------------------
 
-This module contains the propagable for the beta parameter.
+This module contains the propagable for the dispersion parameter.
 """
 from __future__ import annotations
 
@@ -13,10 +13,10 @@ import pandas as pd
 from tfs import TfsDataFrame
 
 from omc3.definitions.optics import OpticsMeasurement
-from omc3.optics_measurements.constants import BETA
+from omc3.optics_measurements.constants import DISPERSION
 from omc3.segment_by_segment import math as sbs_math
 from omc3.segment_by_segment.propagables.abstract import Propagable
-from omc3.segment_by_segment.propagables.phase import Phase
+from omc3.segment_by_segment.propagables.phase import Phase 
 from omc3.segment_by_segment.propagables.utils import PropagableColumns, common_indices
 from omc3.segment_by_segment.segments import SegmentDiffs
 from omc3.utils import logging_tools
@@ -29,24 +29,24 @@ if TYPE_CHECKING:
 LOG = logging_tools.get_logger(__name__)
 
 
-class BetaPhase(Propagable):
+class Dispersion(Propagable):
 
-    _init_pattern = "bet{}_{}"  # format(plane, ini/end)
-    columns: PropagableColumns = PropagableColumns(BETA)
+    _init_pattern = "d{}_{}"  # format(plane, ini/end)
+    columns: PropagableColumns = PropagableColumns(DISPERSION)
 
     @classmethod
     def get_at(cls, names: IndexType, meas: OpticsMeasurement, plane: str) -> ValueErrorType:
         c = cls.columns.planed(plane)
-        beta = meas.beta_phase[plane].loc[names, c.column]
-        error = meas.beta_phase[plane].loc[names, c.error_column]
-        return beta, error
+        dispersion = meas.dispersion[plane].loc[names, c.column]
+        error = meas.dispersion[plane].loc[names, c.error_column]
+        return dispersion, error
     
     @classmethod
     def in_measurement(cls, meas: OpticsMeasurement) -> bool:
-        """ Check if the beta phase is in the measurement data. """
+        """ Check if the dispersion is in the measurement data. """
         try:
-            meas.beta_phase_x
-            meas.beta_phase_y
+            meas.dispersion_x
+            meas.dispersion_y
         except FileNotFoundError:
             return False
         return True
@@ -55,7 +55,7 @@ class BetaPhase(Propagable):
         """ Return the measurement points for the given plane, that are in the segment. """
         return common_indices(
             self.segment_models.forward.index, 
-            self._meas.beta_phase[plane].index
+            self._meas.dispersion[plane].index
         )  
 
     def add_differences(self, segment_diffs: SegmentDiffs):
@@ -63,32 +63,31 @@ class BetaPhase(Propagable):
         dfs = self.get_difference_dataframes()
         for plane, df in dfs.items():
             # save to diffs/write to file (if allow_write is set)
-            segment_diffs.beta_phase[plane] = df
+            segment_diffs.dispersion[plane] = df
 
     def _compute_measured(self, 
             plane: str, 
             seg_model: TfsDataFrame, 
             forward: bool
         ) -> tuple[pd.Series, pd.Series]:
-        """ Compute the beta-beating between the given segment model and the measured values."""
+        """ Compute the dispersion difference between the given segment model and the measured values."""
         init_condition = self._init_start(plane) if forward else self._init_end(plane)
 
         # get the measured values
         names = self.get_segment_observation_points(plane)
-        beta, err_beta = self.get_at(names, self._meas, plane)
+        disp, err_disp = self.get_at(names, self._meas, plane)
 
         # get the propagated values
-        model_beta = seg_model.loc[names, f"{BETA}{plane}"]
+        model_disp = seg_model.loc[names, f"{DISPERSION}{plane}"]
         model_phase = Phase.get_segment_phase(seg_model.loc[names, :], plane, forward) 
 
-        # calculate beta beating
-        beta_beating = (beta - model_beta) / model_beta
+        # calculate difference
+        disp_diff = disp - model_disp
 
         # propagate the error
-        err_beta = err_beta / model_beta
-        propagated_err = sbs_math.propagate_error_beta(model_beta, model_phase, init_condition)
-        total_err = sbs_math.quadratic_add(err_beta, propagated_err)
-        return beta_beating, total_err
+        propagated_err = sbs_math.propagate_error_dispersion(model_disp, model_phase, init_condition)
+        total_err = sbs_math.quadratic_add(err_disp, propagated_err)
+        return disp_diff, total_err
     
     def _compute_correction(
             self,
@@ -97,28 +96,25 @@ class BetaPhase(Propagable):
             seg_model_corr: pd.DataFrame,
             forward: bool,
         ) -> tuple[pd.Series, pd.Series]:
-        """Compute the beta-beating between the nominal and the corrected model."""
+        """Compute the dispersion difference between the nominal and the corrected model."""
         init_condition = self._init_start(plane) if forward else self._init_end(plane)
 
-        # calculate beta beating
-        model_beta = seg_model.loc[:, f"{BETA}{plane}"]
-        corrected_beta = seg_model_corr.loc[:, f"{BETA}{plane}"]
-        beta_beating = (corrected_beta - model_beta) / model_beta
-
+        model_disp = seg_model.loc[:, f"{DISPERSION}{plane}"]
+        corrected_disp = seg_model_corr.loc[:, f"{DISPERSION}{plane}"]
+        disp_diff = corrected_disp - model_disp
+        
         # propagate the error
         model_phase = Phase.get_segment_phase(seg_model, plane, forward)
-        propagated_err = sbs_math.propagate_error_beta(corrected_beta, model_phase, init_condition)
-        return beta_beating, propagated_err
+        propagated_err = sbs_math.propagate_error_dispersion(corrected_disp, model_phase, init_condition)
+        return disp_diff, propagated_err
 
     def _compute_elements(self, plane: str, seg_model: pd.DataFrame, forward: bool):
-        """ Compute get the propagated beta values from the segment model and calculate the propagated error.
-        TODO: Should this be beta-beating with the nominal model (self._elements) ?
-        """
+        """ Compute get the propagated dispersion values from the segment model and calculate the propagated error.  """
         init_condition = self._init_start(plane) if forward else self._init_end(plane)
-        
-        model_beta = seg_model.loc[:, f"{BETA}{plane}"]
+
+        model_disp = seg_model.loc[:, f"{DISPERSION}{plane}"]
 
         # propagate the error
         model_phase = Phase.get_segment_phase(seg_model, plane, forward)
-        propagated_err = sbs_math.propagate_error_beta(model_beta, model_phase, init_condition)
-        return model_beta, propagated_err
+        propagated_err = sbs_math.propagate_error_dispersion(model_disp, model_phase, init_condition)
+        return model_disp, propagated_err