+++ omc3/segment_by_segment/segments.py
@@ -9,9 +9,11 @@ from __future__ import annotations
 from dataclasses import dataclass
 from pathlib import Path
 
+from tfs import TfsDataFrame
 from tfs.collection import Tfs, TfsCollection
 
 from omc3.optics_measurements.constants import (
+    ALPHA_NAME,
     AMP_BETA_NAME,
     BETA_NAME,
     DISPERSION_NAME,
@@ -19,6 +21,8 @@ from omc3.optics_measurements.constants import (
     KMOD_BETA_NAME,
     NORM_DISP_NAME,
     PHASE_NAME,
+    F1001_NAME,
+    F1010_NAME,
 )
 from omc3.segment_by_segment.constants import (
     TWISS_BACKWARD,
@@ -80,10 +84,10 @@ class SegmentModels(TfsCollection):
         directory: The path where to find the models.
         segment: A segment instance corresponding to the model to load.
     """
-    forward = Tfs(TWISS_FORWARD, two_planes=False)
-    backward = Tfs(TWISS_BACKWARD, two_planes=False)
-    forward_corrected = Tfs(TWISS_FORWARD_CORRECTED, two_planes=False)
-    backward_corrected = Tfs(TWISS_BACKWARD_CORRECTED, two_planes=False)
+    forward: TfsDataFrame = Tfs(TWISS_FORWARD, two_planes=False)   # whenever used, these are TfsDataFrames
+    backward: TfsDataFrame = Tfs(TWISS_BACKWARD, two_planes=False)
+    forward_corrected: TfsDataFrame = Tfs(TWISS_FORWARD_CORRECTED, two_planes=False)
+    backward_corrected: TfsDataFrame = Tfs(TWISS_BACKWARD_CORRECTED, two_planes=False)
 
     def __init__(self, directory: Path, segment: Segment):
         super(SegmentModels, self).__init__(directory)
@@ -106,17 +110,19 @@ class SegmentDiffs(TfsCollection):
 
     phase = Tfs(f"{PREFIX}{PHASE_NAME}{{plane}}_{{name}}{EXT}")
     beta_phase = Tfs(f"{PREFIX}{BETA_NAME}{{plane}}_{{name}}{EXT}")
+    alpha_phase = Tfs(f"{PREFIX}{ALPHA_NAME}{{plane}}_{{name}}{EXT}")
     beta_kmod = Tfs(f"{PREFIX}{KMOD_BETA_NAME}{{plane}}_{{name}}{EXT}")
     beta_amp = Tfs(f"{PREFIX}{AMP_BETA_NAME}{{plane}}_{{name}}{EXT}")
     dispersion = Tfs(f"{PREFIX}{DISPERSION_NAME}{{plane}}_{{name}}{EXT}")
     norm_dispersion = Tfs(f"{PREFIX}{NORM_DISP_NAME}{{plane}}_{{name}}{EXT}")
-    # TODO: Add coupling!
+    f1001 = Tfs(f"{PREFIX}{F1001_NAME}_{{name}}{EXT}", two_planes=False)
+    f1010 = Tfs(f"{PREFIX}{F1010_NAME}_{{name}}{EXT}", two_planes=False)
 
     def __init__(self, directory: Path, segment_name: str, *args, **kwargs):
         super(SegmentDiffs, self).__init__(directory, *args, **kwargs)
         self.segment_name = segment_name 
 
-    def _get_filename(self, template: str, plane: str=None):
+    def _get_filename(self, template: str, plane: str | None = None):
         if plane is None:
             return template.format(name=self.segment_name)
         return template.format(plane=plane.lower(), name=self.segment_name)